requirements.txt
python-dotenv
ipykernel
ipywidgets
tilemapbase
geopandas
pandas
numpy
matplotlib
geopy
neo4j
tqdm
pydantic
requests
llama-index-core
llama-index-llms-openai
httpx
Jinja2
scikit-learn
tiktoken
bert-score
nltk
overpass